Display validator

+++ Test/src/main/java/ua/com/shop/dao/DisplaySizeDao.java
@@ -8,6 +8,6 @@ import ua.com.shop.entity.DisplaySize;
 public interface DisplaySizeDao extends JpaRepository<DisplaySize, Integer> {
 
 	@Query("SELECT d FROM DisplaySize d WHERE d.size = ?1")
-	DisplaySize findUnique(int size);
+	DisplaySize findUnique(double size);
 
 }
